+++ packages/expo-video/src/NativeVideoModule.ts
@@ -5,6 +5,9 @@ import type { VideoPlayer } from './VideoPlayer.types';
 type ExpoVideoModule = {
   VideoPlayer: typeof VideoPlayer;
   isPictureInPictureSupported(): boolean;
+  setVideoCacheSizeAsync(sizeBytes: number): Promise<void>;
+  cleanVideoCacheAsync(): Promise<void>;
+  getCurrentVideoCacheSize(): number;
 };
 
 export default requireNativeModule<ExpoVideoModule>('ExpoVideo');