+++ packages/expo-video/src/VideoView.tsx
@@ -5,16 +5,6 @@ import NativeVideoView from './NativeVideoView';
 import type { VideoPlayer } from './VideoPlayer.types';
 import type { VideoViewProps } from './VideoView.types';
 
-/**
- * Returns whether the current device supports Picture in Picture (PiP) mode.
- * @returns A `boolean` which is `true` if the device supports PiP mode, and `false` otherwise.
- * @platform android
- * @platform ios
- */
-export function isPictureInPictureSupported(): boolean {
-  return NativeVideoModule.isPictureInPictureSupported();
-}
-
 export class VideoView extends PureComponent<VideoViewProps> {
   nativeRef = createRef<any>();